replit/pyproject.toml
[tool.poetry]
name = "repl-nix-audiotranscriptionandblogposter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
openai = "^1.44.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pyjwt = "^2.9.0"
gtts = "^2.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
